Cleanup example script
