+++ examples/demo/run-transient.py
@@ -6,13 +6,12 @@ output_dir = "."
 pvd_out = None
 restart = None
 checkpoint = "checkpoint.h5"
-# checkpoint = None
 
 flow = hgym.Cylinder(Re=100, restart=restart, mesh="medium")
 
 # Time step
 Tf = 300
-dt = 0.01
+dt = 0.1
 
 
 def log_postprocess(flow):
@@ -26,15 +25,15 @@ print_fmt = "t: {0:0.2f},\t\t CL: {1:0.3f},\t\t CD: {2:0.03f}\t\t Mem: {3:0.1f}"
 log = hgym.utils.io.LogCallback(
     postprocess=log_postprocess,
     nvals=3,
-    interval=10,
+    interval=1,
     print_fmt=print_fmt,
     filename=None,
 )
 
-# callbacks = [log, hgym.utils.io.CheckpointCallback(interval=100, filename=checkpoint)]
-callbacks = [
-    log,
-]
+callbacks = [log, hgym.utils.io.CheckpointCallback(interval=10, filename=checkpoint)]
+# callbacks = [
+#     log,
+# ]
 
 
 def controller(t, y):
@@ -48,5 +47,5 @@ hgym.integrate(
     dt=dt,
     callbacks=callbacks,
     method="BDF",
-    controller=controller,
+    # controller=controller,
 )
